Added LwIP support for stm32-cmake
This feature allows using the LwIP source code and OS abstraction layer
provided by STM to build networking applications for STM MCUs with
stm32-cmake.

It uses the CMakeLists.txt already contained in the LwIP sources.
It is difficult to provide examples for LwIP code for various reasons:

1. Networking code often requires special code which is board and
   architecture specific. Examples: PHY or linker scripts for special
   lwip sections

2. STM licenses a lot of their example code differently than the rest of
   the HAL repositories. This license might be incompatible to the MIT
   license.

A separate project is linked in the README example section for a basic
UDP echoserver on the H743ZI board.
The README was updated with intructions on how to use the LwIP support.