Add excluding drivers from autodetection
By default, the list of drivers available for a family is deduced from
the `.c` files which are present.  However, some drivers appear to have
been included by accident and are broken (e.g. the `MDIOS` drivers for
the MP1 family

Add:
- Exclude lists for drivers (`EXCLUDED_HAL_DRIVERS_<FAMILY>`)
- The ability for users to override this exclude list if required
  (through `EXCLUDED_HAL_DRIVERS_<FAMILY>_OVERRIDE`)
    - Seting `EXCLUDED_HAL_DRIVERS_<FAMILY>_OVERRIDE` to the empty
      string will force all modules to be enabled.

Also, exclude the `MDIOS` drivers for STM32 MP1 by default.  The HAL
driver for this IP is broken in the current release (1.6.0)