Add '--default-script' for SDK provided linker script
Previously, when using `stm32_add_linker_script` to define a custom
linker script, the SDK provided default linker script, that used the
`-T/--script` flag, that led to conflicts when you wanted a complete
new user provided linker script.

The `--default-script` flag changes this behaviour and makes it possible
to derive from the SDK provided linker script, but also enables the
possibility to create a completely new linker script.

This commit adds a new internal command `stm32_add_default_linker_script`
that is used internal for the `--default-script` flag, and the normal
users shouldn't be affected by this. Thus this should not be a breaking
change if this function is used as normally intended.